backend/python-implementations/requirements_official.txt
# Official CopilotKit Python SDK
copilotkit==0.1.72

# LangGraph for agent
langgraph>=0.6.6
langchain-core>=0.3.0
langchain-openai>=0.2.0

# FastAPI
fastapi>=0.115.0
uvicorn>=0.32.0

# Environment
python-dotenv>=1.0.0

